refactor(controllers): extract LaunchNotifications from HelpSessionController
- Move notifyLaunchFailed, notifyAssistantServicesUnavailable,
  notifyInstallCorrupted, launch-error banner/toast routing, and the
  resume/snapshot banner auto-dismiss timers into a sibling module
- Controller delegates through a narrow host interface (patch/isPanelOpen)
  — no behavior change
